refactor: move usePrefetchSignal into signal file
